+++ NEWS.md
@@ -1,5 +1,6 @@
 # medfateland 2.7.1
 - Subwatersheds returned as part of spatial output in *_land simulations
+- Bug correction for .check_meteo_column_input issue #5.
 
 # medfateland 2.7.0
 - Runoff-runon dynamics coupled to local water balance in TETIS