+++ NEWS.md
@@ -1,6 +1,8 @@
 # medfateland 2.7.1
 - Subwatersheds returned as part of spatial output in *_land simulations
-- Bug correction for .check_meteo_column_input issue #5.
+- Bug correction for .check_meteo_column_input issue #5
+- Daily substeps available for base flows
+- New plot option for watershed-level simulation results
 
 # medfateland 2.7.0
 - Runoff-runon dynamics coupled to local water balance in TETIS